Build, install and use cloog-isl libs for android targets:
build from source and install to sysroot of toolchain,
then add the libs to LD_LIBRARY_PATH and LIBRARY_PATH
to compile the toolchains with.  This elimates the need to
install any cloog or isl packages in order to compile a sabermod
toolchain or a android ROM with a sabermod toolchain.

Signed-off-by: Paul Beeler <pbeeler80@gmail.com>